requirements.txt
Flask==1.1.2
gunicorn==19.5.0
flask-cors
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyopenssl
